+++ src/main/java/com/kosa/mini/api/entity/ReviewReply.java
@@ -15,7 +15,7 @@ public class ReviewReply {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "reply_id")
-    private Long replyId;
+    private Integer replyId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "review_id", nullable = false)
